+++ trend-collector/trendcollector/libs/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env /Users/yasui/Practice/trend_collector/trend-collector/venv/bin/python
 
 from .services.client import *
-from .twitter_v2 import *
+from .infrastractures.twitter_v2 import *
 from .conf import *
 from .response import *
 from .infrastractures import *
