+++ trend-collector/trendcollector/server.py
@@ -9,9 +9,9 @@ from typing import Union
 
 from libs import Base
 from libs.infrastractures import TrendRepository, TwitterAccountRepository
+from libs.infrastractures.twitter_v2 import TwitterV2
 from libs.response import *
 from libs.services.collector import TwitterCollector
-from libs.twitter_v2 import TwitterV2
 
 logger = logging.getLogger('uvicorn')
 env = Environment()
